skip duplicate locales when normalized
If you try to generate "en_US.UTF-8 UTF-8" and "en_US.UTF8 UTF-8", we
end up generating the locale twice even though it normalizes down to
the same value (or, in the case of --update, we skip the 1st and then
always generate the 2nd).  Keep track of all locales we've processed
so we can dedupe as we go.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>