Only check the prospective archive dir once there is work to be done
The locale-gen(1) utility combines the values of the 'prefix' and
'locale_dir' variables to form the directory that is expected to contain
the locale archive, then checks whether said directory exists and can be
modified by the effective UID. Presently, this check is performed
immediately after the command line options are validated.

Instead, perform the check at the point that locale generation is about
to commence. Further, move the code responsible for performing the check
into a new subroutine by the name of check_archive_dir.

Signed-off-by: Kerin Millar <kfm@plushkava.net>