Render compatible with the compile-locales USE flag of sys-libs/glibc
Version 3.0 of locale-gen(1) fails in the case that it is invoked by the
sys-libs/glibc ebuild, provided that the "compile-locales" USE flag is
in effect. It was developed under the understanding that the flag might
eventually be removed and unexpectedly committed to the repo before any
meaningful testing had been performed with the flag enabled.

The particulars of the two modes of failure are described herewith.

Firstly, the ebuild executes locale-gen(1) with "${ED}" as a prefix.

  locale-gen --prefix /var/tmp/portage/sys-libs/glibc-2.41-r4/image

Secondly, the program declares the $prefix, $locale_dir and
$gentoo_prefix variables. Given the above value of the --prefix option,
the values of these variables shall be as follows.

  $prefix = "var/tmp/portage/sys-libs/glibc-2.41-r4/image"
  $locale_dir = "/usr/lib/locale"
  $gentoo_prefix = ""

Thirdly, having compiled the requested locales, the program calls the
generate_archive() subroutine, passing the values of the $prefix and
$locale_dir variables (but not the $gentoo_prefix variable).

Fourthly, the generate_archive() subroutine declares the $output_dir
variable.

  # Assigns "./var/tmp/portage/sys-libs/glibc-2.41-r4/image/usr/lib/locale"
  my $output_dir = catdir('.', $prefix, $locale_dir)

Fifthly, the generate_archive subroutine executes mkdir(1), so as to
create the directory specified by $output_dir.

  mkdir -p ./var/tmp/portage/sys-libs/glibc-2.41-r4/image/usr/lib/locale

Sixthly, the localedef(1) utility is executed with "." as its first
argument. Internally, it joins "." to "/usr/lib/locale" and attempts to
write a new archive there. This is the point at which a failure occurs,
for the "./usr/lib/locale" directory doesn't exist. In turn, that is
because the wrong value was assigned to the $output_dir variable.

Address this issue by also passing the value of the $gentoo_prefix
variable to the generate_archive() subroutine, and using that to form
the value of $output_dir, like so.

  # Correctly assigns "./usr/lib/locale"
  my $output_dir = catdir('.', $gentoo_prefix, $locale_dir)

This works as intended in Gentoo Prefix environments. Imagine that such
an environment is situated at "/home/gentoo/gentoo". The values of the
$prefix, $locale_dir and $gentoo_prefix variables shall be as follows.

  $prefix = "/home/gentoo/gentoo/var/tmp/portage/sys-libs/glibc-2.41-r4/image/home/gentoo/gentoo"
  $locale_dir = "/usr/lib/locale"
  $gentoo_prefix = "/home/gentoo/gentoo"

Consequently, $output_dir shall be formed in the following way.

  # Correctly assigns "./home/gentoo/gentoo/usr/lib/locale"
  my $output_dir = catdir('.', $gentoo_prefix, $locale_dir)

Such is correct because the location of the Gentoo Prefix environment is
integrated into the localedef(1) binary. That is, one cannot simply
specify the desired output directory to localedef(1) in full.

Additionally, the parse_opts() subroutine checks whether the value of
the --prefix option is equal to the path of the automatically detected
Gentoo Prefix, if any. This causes a failure in its own right since the
value of $prefix cannot possibly be equal to $gentoo_prefix in the case
that the "compile-locales" flag is in effect. Jettison this check.

Signed-off-by: Kerin Millar <kfm@plushkava.net>