Attempt to create the archive directory, if missing
Have the check_archive_dir() subroutine quietly attempt to create the
directory that is supposed to contain the locale archive, provided that
it is missing. Versions prior to 3.0 would also attempt to create the
directory and this seems to me to be a reasonable thing to do, for it
relieves the sys-libs/glibc ebuild of the burden of guaranteeing that
the directory exists for all its modes of operation.

See-also: 02b2ac6ee591e4ee65eb131fa8038dc473678527
Signed-off-by: Kerin Millar <kfm@plushkava.net>