locale-gen.8: rewrite in the style that a technical writer might
This commit improves the locale-gen(8) manual by exacting a style of
English that is characteristic of technical documentation, by which I
mean documentation that has been written by a technical writer. The
DESCRIPTION section has undergone the most notable changes, better
describing the purpose and behaviour of the program. The OPTIONS section
has also been revised, and the FILES section now contains entries for
the "SUPPORTED" and "locale-archive" files.

Bug: https://bugs.gentoo.org/945269
Signed-off-by: Kerin Millar <kfm@plushkava.net>