locale.gen: show only UTF-8 based locales as commented examples
Presently, the default locale.gen file contains only 14 (commented)
locale definitions. Of the 9 countries covered, 6 are lacking any
definitions that incorporate the UTF-8 character map. This makes no
sense. As of 2025, the overwhelming majority of users ought to be using
the UTF-8 character map, along with the Unicode Collation Algorithm. In
no way should Gentoo be encouraging new users not to do so, especially
given that some of them would be hard pressed to tell a locale apart
from a latrine.

Address this matter by replacing the aforementioned definitions with all
325 of the locales from the /usr/share/i18n/SUPPORTED file that specify
the UTF-8 character map (not counting C.UTF-8). Also, present them as a
columnated list, rendering it considerably more pleasant to read.

Signed-off-by: Kerin Millar <kfm@plushkava.net>