Improve the diagnostic where /usr/share/i18n/locales cannot be entered
Consider a scenario in which locale-gen(8) is unable to switch to the
directory containing the glibc locale definition files.

# locale-gen --prefix /var/empty
sh: 1: cd: can't cd to /var/empty/usr/share/i18n/locales
locale-gen: Failed to compose a list of valid locale names from '/var/e
pty/usr/share/i18n/locales'

In this case, the implementation of sh(1) is provided by dash. I find
its diagnostic message to be lacking because it does not explain why it
cannot switch to the requested directory. That is, it could have
reported the chdir(2) error.

Address this minor issue by having the Perl interpreter attempt to
switch to the directory before executing sh(1) and find(1) and throwing
a tailored exception if it cannot.

# locale-gen --prefix /var/empty
locale-gen: Can't chdir to '/var/empty/usr/share/i18n/locales': No such
file or directory

Signed-off-by: Kerin Millar <kfm@plushkava.net>