Require either a null string or an absolute path for --prefix
The --prefix option is only intended for use by the sys-libs/glibc
ebuild, which is expected to pass either the value of ED or EROOT. For
the --prefix option-argument to specify something other than either a
null string or an absolute path would imply that something has gone
catastrophically wrong. Reject the option-argument in that case.

Signed-off-by: Kerin Millar <kfm@plushkava.net>