Claim that glibc is absent where localedef(1) output cannot be parsed
Presently, the get_locale_dir() subroutine executes localedef(1) with
the --help option in order to capture the locale path. Should it be
unable to do so, it throws an exception that's intended for developers
rather than end-users. The experience can thus be an ugly one in the
case that locale-gen(8) is run on a foreign platform.

# locale-gen
localedef: illegal option -- -
Usage: localedef [options] localename
[options] are:
  -D          : BSD-style output
  -b          : big-endian output
  -c          : ignore warnings
  -l          : little-endian output
  -v          : verbose output
  -U          : ignore undefined symbols
  -f charmap  : use given charmap file
  -u encoding : assume encoding
  -w widths   : use screen widths file
  -i locsrc   : source file for locale
  -V version  : version string for locale
Can't parse locale directory from localedef(1) output:  at ./locale-gen line 130.

Address this minor issue in a twofold manner. Firstly, suppress the
standard error of localedef(1). Secondly, claim that the operating
system does not use glibc in either of the following cases.

a) localedef(1) exits with a status of 0 and its output cannot be parsed
b) localedef(1) or sh(1) otherwise exits normally, with any status

The new diagnostic is depicted beneath.

# locale-gen
locale-gen: Aborting because the OS does not appear to use GNU libc

Signed-off-by: Kerin Millar <kfm@plushkava.net>