locale.gen: remove the comments masquerading as documentation
It is not worth the effort to revise them in accordance with the current
state of the locale.gen(5) man page. Besides which, it is hardly
untoward to expect a user to enter a simple command in order to read the
documentation.

See-also: fe5bd5fa7bb7a9aa231f5fdd424f221884f505a0
Signed-off-by: Kerin Millar <kfm@plushkava.net>