Coerce LC_ALL=C for localedef(1) in get_locale_dir()
The GNU implementation of localedef(1) has translations that affect the
output produced by the --help option.

# LC_MESSAGES=ja_JP.UTF-8 localedef --help | grep -B1 ^/usr/lib/locale
ロケールパス用のシステムディレクトリ       :
/usr/lib/locale:/usr/share/i18n

Presently, the get_locale_dir() subroutine does not account for this and
is therefore unable to match against "locale dir", after which
locale-gen(8) erroneously claims that the OS isn't using glibc.

Address this issue by coercing the C (POSIX) locale for that particular
invocation.

Closes: https://bugs.gentoo.org/961219
Reported-by: YOSHIOKA Takuma <lo48576@hard-wi.red>
Signed-off-by: Kerin Millar <kfm@plushkava.net>