locale.gen.5: Correct the locale name synopsis per XPG Issue 4
Currently, Section 7.6 of the glibc manual states that "most locale
names follow XPG syntax". In characteristic GNU fashion, it fails to
explain what XPG is or precisely which issue of the specification it is
referring to. It goes on to present the following locale name synopsis.

  language[_territory[.codeset]][@modifier]

Such is a valid synopsis for XPG3. But glibc does not observe the
conventions of XPG3. Instead, it observes the conventions of XPG4. That
is, the X/Open Common Applications Environment (CAE) Portability Guide
Issue 4 (XPG4) specification. As such, the synopsis should actually be
written as follows.

  language[_territory][.codeset][@modifier]

The difference is that all but the language portion of a locale name are
rendered optional by Issue 4. Indeed, the glibc manual goes on to
literally state that, "besides the first part, all of them are allowed
to be missing."

Though I cannot correct the glibc manual directly, I can correct the
locale.gen(5) man page. Do so by having it show the XPG4 synopsis.

Link: https://sourceware.org/glibc/manual/latest/html_node/Locale-Names.html
Signed-off-by: Kerin Millar <kfm@plushkava.net>