Eliminate all unescaped line continuations from the man pages
Presently, some of the paragraphs in the man pages are wrapped at the 80
column mark. However, the effect of an unescaped newline is significant.
As groff(7) states, a newline "puts an inter-word space onto the output
and, if filling is enabled, triggers end-of-sentence recognition on the
preceding text."

Address this issue by observing the status quo and refraining from
wrapping the affected text.

Signed-off-by: Kerin Millar <kfm@plushkava.net>