locale.gen: refine the annotations
Adjust the comments for the mnw_MM and shn_MM locales so that Myanmar is
referred to as "Myanmar/Burma", thereby matching the my_MM locale.

Adjust the comments for the ku_TR and tr_TR locales so that Turkey is
referred to by its English name rather than as "Türkiye". All of the
other entries employ English names.

Signed-off-by: Kerin Millar <kfm@plushkava.net>