Report the size of the installed archive
Upon success, have the final diagnostic message report the size of the
installed archive. Further, have it accurately report how many locales
are incorporated by the archive for the -u option.

# printf '%s UTF-8\n' en_US en_GB | locale-gen -c - | tail -n2
Adding 3 locales to the locale archive ...
Successfully installed an archive containing 3 locales, of 2.93 MiB in size.

# printf '%s UTF-8\n' en_AU | locale-gen -u -c - | tail -n2
Adding 1 locale to the locale archive ...
Successfully installed an archive containing 4 locales, of 2.94 MiB in size.

Signed-off-by: Kerin Millar <kfm@plushkava.net>