Use ($CHILD_ERROR & 0x7F) to check for death by signal
Should a child process die to a signal, the lower 7 bits of the $?
variable shall contain the signal number, whereas the 8th bit shall
indicate whether a core dump was produced. Hence, either of ($? & 0xFF)
and ($? & 0x7F) can be used to determine whether a process exited
normally. The latter expression is idiomatic, however.

Signed-off-by: Kerin Millar <kfm@plushkava.net>