Better align the behaviour of normalize() with localedef(1)
The purpose of the normalize() subroutine is to normalize the codeset
portion of a locale name as the localedef(1) utility would; at least, to
the extent that is necessary for locale-gen(8) to function correctly.

Presently, it does so by splitting the string into two parts (with the
<period> character serving as a separator), stripping all instances of
the <hyphen-minus> from the second part, then converting the second part
to lowercase. This approach dates from April 2006, when the --update
option was introduced. However, it is not strictly correct. For one
thing, only the codeset part is supposed to be altered. Consider the
following locale name.

  de_DE.ISO-8859-15@euro

The present routine operates on a substring of "ISO-8859-15@euro".
Instead, it ought to operate only on "ISO-8859-15". While it seems
unlikely ever to occur, imagine a scenario in which GNU introduces a new
modifier that does not consist exclusively of lower case characters.

Another issue is that the routine strips only the <hyphen-minus>
character. Instead, it ought to strip all non-alphanumeric characters,
as does the normalize_codeset() function of the localedef(1) utility.

Revise the normalize() subroutine so as to address both of these issues
and maintain a theoretically high degree of forward-compatibility.

Fixes: 2df969d53ea596038a4857060a42d7f2fd25d7e3
Link: https://sourceware.org/git?p=glibc.git;a=blob;f=locale/programs/localedef.c;hb=glibc-2.42#l561
Signed-off-by: Kerin Millar <kfm@plushkava.net>