Don't call parse_config() from a try block in read_config()
Presently, the read_config() subroutine employs a try block to catch
exceptions thrown in the case that a given config file cannot be opened.
However, it also calls the parse_config() subroutine from that same
block. Given that config parsing exceptions need not be caught - even if
only to be re-thrown - there is no sense in doing so. Hoist the call so
that it is no longer enclosed by the try block.

Signed-off-by: Kerin Millar <kfm@plushkava.net>