Introduce the mkconfig utility
The mkconfig utility writes a sample /etc/locale.gen file to standard
output as a columnated list, consisting only of UTF-8 locales supported
by the installed version of glibc, with comments indicating the
languages and territories in plain English.

Given the necessary effort to integrate it into the sys-apps/locale-gen
ebuild, it could be used by the next release of locale-gen. The idea is
not to have to carry a stock "locale.gen" file that requires manual
review upon each new glibc release.

Signed-off-by: Kerin Millar <kfm@plushkava.net>