mkconfig: don't transliterate paths in map_locale_attributes()
Presently, the map_locale_attributes() subroutine employs a two-stage
pipeline, where grep(1) is used to find lines defining the "language"
and "territory" attributes, with iconv(1) transliterating to US-ASCII.
This is done so as to eliminate diacritics and produce a config that is,
at once, valid US-ASCII and valid UTF-8. However, there is a potential
issue with this approach. Imagine a scenario in which Gentoo is
installed beneath the following prefix.

  /home/gentoo/gentøø-linux

In that case, the pathname will be transliterated to:

  /home/gentoo/gentoo-linux

Thus, the regular expression that separates the pathname from the line
will fail to match. At first, I contemplated switching to the
/usr/share/i18n/locales directory before grep(1) is executed. To do so
would almost certainly have sufficed, because the names of the files
residing in that directory are unaffected by transliteration, and this
will remain the case for the foreseeable future. Then I realised that
Perl is able to perform transliteration using only core modules.

As such, address this issue by having Perl be directly responsible for
opening and parsing the locale files. For any field whose value does not
consist entirely of bytes in the 0x00 - 0x7F range, transliterate the
diacritics in the following way.

1) Decode the bytes to characters (presuming UTF-8 as the encoding)
2) Convert the characters to the NFKD normal form
3) Strip non-spacing combining marks with a zero advance width

This approach appears adequate, and remains about as fast.

It should be noted that dev-perl/File-Slurper is now a requirement. I
favour this module because its read_binary() subroutine performs an
extremely fast unbuffered 'slurp'. Should this requirement be considered
problematic, it may be removed before the next release is issued.

Signed-off-by: Kerin Millar <kfm@plushkava.net>