mkconfig: don't operate on refs of $language and $territory
Presently, the map_locale_attributes() subroutine creates references of
the $language and $territory variables, so as to be able to loop over -
and operate on - both. Yet one need not bother, since the for keyword
renders the topicaliser a temporary alias of each item in the list.

See-also: 87b724f3b5bf14ade06d3a15123e08677d05efbe
Signed-off-by: Kerin Millar <kfm@plushkava.net>