Reduce the extent to which render_printable() is used
Presently, there are diagnostic messages in which pathnames are enclosed
by single quotes, and others that use the render_printable() subroutine
to convert to ASCII-constrained JSON, duly enclosing them within double
quotes. Partly address this inconsistency by refraining from calling
render_printable() other than in the following three cases.

1) throwing from normalize() for malformed locale names
2) throwing from read_config() where no locale entries were found
3) throwing from parse_config()

There remains an inconsistency in that the second of these cases
continues to render pathnames as JSON. I have let it be because it is
useful to be able to render the entirety of the @paths array in an
unambiguous fashion.

Signed-off-by: Kerin Millar <kfm@plushkava.net>