Add support for preserving/restoring SELinux contexts
Among the changes brought by locale-gen v3.0 was the following:

- During the archive generation phase, localedef(1) is made to generate
  its archive in a temporary directory. Only if it succeeds shall any
  existing archive be replaced, with locale-gen(1) now assuming that
  responsibility. The existing archive is guaranteed to be replaced
  atomically or not at all.

Such is problematic for operating environments in which SELinux is
enabled because the newly created archive ends up lacking the
appropriate label. Consequently, programs are unable to read the archive
until such time as the user restores the label.

Address this issue in a manner twofold. Firstly, in the event that a
prior archive exists, execute the chcon(1) utility so as to copy the
label from the old archive to the new archive, just before the former is
replaced by a rename(2) syscall. Secondly, in the event that a prior
archive does not exist, execute the restorecon(8) utility in an attempt
to apply the appropriate label.

No labelling shall be attempted unless the underlying filesystem is
found to be mounted with the "seclabel" option in effect. Further,
restorecon(8) shall only be executed if is found in PATH and the
--prefix option was either a) unspecified b) specified as "/".

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Tested-by: Nicolas PARLANT <nicolas.parlant@parhuet.fr>
Tested-by: Kenton Groombridge <concord@gentoo.org>
Closes: https://bugs.gentoo.org/962753