Delete interim archives for which installation fails
Presently, the generate_archive() subroutine performs these steps:

1) generates a new locale archive beneath $TEMPDIR
2) installs it with a suffix that incorporates $$ (the PID of perl)
3) calls rename so as to conclude the installation procedure

However, it is possible for an error to occur after the second step has
successfully concluded, yet before the third step has successfully
concluded. In that case, the interim archive will linger.

# ls /usr/lib/locale
locale-archive locale-archive.3241796

Address this issue by replacing the $TEMPDIR scalar variable with the
@TEMPFILES array variable and ensuring that the path of the interim
archive is pushed to it. The elements of the array shall be purged by
the rm(1) utility, as executed by the END block.

Link: https://bugs.gentoo.org/962753#c9
Signed-off-by: Kerin Millar <kfm@plushkava.net>