Ignore findmnt(8) exiting 1 where / is not a mountpoint
Presently, the has_mount_option() subroutine executes findmnt(8) so as
to determine the mount options of the mountpoint subsuming the directory
that is intended to contain the locale archive.

However, in some cases, such as where using qemu-user, no / mountpoint
exists. In that case, it is highly probable that findmnt -T will recurse
upwards until it reaches /, only to exit 1 because it cannot be
resolved.

Address this issue by disregarding an exit status of 1 in the case that
/ is found not to be a mountpoint.

Bug: https://bugs.gentoo.org/962817
Bug: https://bugs.gentoo.org/962753
Signed-off-by: Kerin Millar <kfm@plushkava.net>