Capture STDERR with a piped open in generate_archive()
Presently, the generate_archive() subroutine calls the redirect_stderr()
subroutine. The latter manipulates file descriptor #2 in the initial
perl process before executing a callback that causes perl to fork and
execute localedef(1) for the purpose of building the locale archive.
Further, the file descriptor in question is re-opened with a file in
read/write mode. Once localedef(1) exits, the file offset is reset to 0
so that the captured errors may be read.

Simplify matters by initially forking perl with the open builtin and
having the child process perform the equivalent of dup2(1, 2) before
proceeding to exec localedef(1). That way, the parent process can read
directly from the resulting pipe and there is no need to restore the
prior state of the file descriptor.

Signed-off-by: Kerin Millar <kfm@plushkava.net>