Have list_locales() call close to wait for the child
Presently, the list_locales() subroutine uses the waitpid builtin to
wait for the child process in which localedef(1) is executed. Instead,
have the parent process close its end of the pipe. Doing so suffices
because the closure of a filehandle produced by a piped open returns
false if one of the other syscalls involved fails or if the process
exits with a non-zero status. In other words, the child shall be
implicitly waited for.

Besides, it is sensible to close a pipe upon the very moment it no
longer serves any purpose.

See-also: a9c2999d4804ecb5b97782f5011b04158b72317a
Signed-off-by: Kerin Millar <kfm@plushkava.net>