Validate the config strictly against the SUPPORTED file
Presently, the locale.gen configuration file supports arbitrary
locale/charmap pairings, provided that the specified locale and charmap
are both known to glibc.

This commit alters the behaviour so that only the combinations specified
by the /usr/share/i18n/SUPPORTED file are tolerated.

# echo 'en_US BIG5' | locale-gen -c -
locale-gen: Unsupported locale/charmap combination at /dev/stdin[1]: "en_US BIG5"

The reasons for this change are twofold. Firstly, to dispense with any
bickering as to precisely what should constitute a valid configuration.
Indeed, the new behaviour accords with the text of locale.gen(5), which
directly states that the supported combinations are to be found within
the SUPPORTED file, which - lest we forget - is a component of glibc.

Secondly, to lay the groundwork for an ensuing commit, which shall
pertain to a matter of backward-compatibility.

Bug: https://bugs.gentoo.org/963974
Signed-off-by: Kerin Millar <kfm@plushkava.net>