Produce short-form aliases for a subset of valid locales
Where localedef(1) is instructed to incorporate a directory into an
archive, it is the name of the directory that determines the apparent
name of the locale. However, if the name of the directory does not
incorporate a codeset part, localedef(1) will sniff the codeset from the
contents of the directory then proceed to incorporate two locales, with
one serving as an alias of the other.

# cd /usr/lib/locale
# localedef --no-archive -i en_US -f ISO-8859-1 en_US # ambiguous
# localedef --add-to-archive --replace en_US
# locale -a
en_US
en_US.iso88591

These short-form aliases are bad for interoperability. Even among
GNU/Linux vendors, there is no consensus as to the circumstances under
which they should be offered (if at all), nor the character maps that
should be implied by effecting any of them.

That being said, locale-gen-2.23 would incorporate these aliases under
certain conditions, whereas >=locale-gen-3.0 currently does not. As
such, there is a legitimate concern as to backward-compatibility. It
stands to reason that some Gentoo users will have chosen to effect a
short-form locale. For these locales to disappear can have adverse
effects on any applications and services that they depend upon.

Address this concern by having locale-gen(8) predictably incorporate
short-form aliases for certain locale/charmap combinations. Consider the
following configuration.

# All supported combinations for the USA
en_US.UTF-8    UTF-8
en_US          ISO-8859-1 # this will be aliased as "en_US"

# All supported combinations for China
zh_CN.GB18030  GB18030
zh_CN.GBK      GBK
zh_CN.UTF-8    UTF-8
zh_CN          GB2312 # this will be aliased as "zh_CN"

This is also the strategy employed by (at least) Fedora and Void Linux.

Reported-by: Horste Prote <prote@fmi.uni-stuttgart.de>
Closes: https://bugs.gentoo.org/963974
Signed-off-by: Kerin Millar <kfm@plushkava.net>