mkconfig: better detect missing/incomplete locales
In the case that the map_locale_attributes() subroutine encounters a
locale for which it is unable to discern its language and/or territory
attributes, incorporate it into the 'attr_by' hash as an empty hash
reference. That way, it becomes trivial to distinguish between the case
where the locale is missing (for its key will not exist) and the case
where no language and/or territory attribute could be extracted from it
(for its value will be an empty hashref).

Signed-off-by: Kerin Millar <kfm@plushkava.net>