Allow localename fields to forgo a (redundant) codeset
Presently, the locale.gen(5) configuration file is validated strictly
against the contents of the SUPPORTED file. Consequently, for any given
pair of localename and charmap fields, it may or may not be necessary to
redundantly incorporate the charmap into the localename field.

# Valid. Note that the the charmap must be expressed twice.
en_US.UTF-8 UTF-8

# Valid.
en_US ISO-8559-1

# Invalid because there is no precise match against a SUPPORTED entry.
en_US UTF_8

This commit adjusts the locale.gen(5) validation methodology in such a
way that it is never necessary to incorporate the charmap into the
localename field. Backward-compatibility is maintained. That is, any
configuration that could be validated by locale-gen 3.9 shall continue
to be valid.

To understand how this is accomplished, consider the prior behaviour of
the map_supported_combinations() subroutine. Imagine that is tasked with
reading the two entries in the SUPPORTED file that pertain to the United
States. In that case, it would generate the following data structure.

{
   "en_US": {
      "ISO-8859-1": 1, // allows "en_US ISO-8859-1" with "en_US" alias
   },
   "en_US.UTF-8": {
      "UTF-8": 1       // allows "en_US.UTF-8 UTF-8"
   }
}

Now, the data structure shall instead be as follows.

{
   "en_US": {
      "ISO-8859-1": 1, // allows "en_US ISO-8859-1" with "en_US" alias
      "UTF-8": 0       // allows "en_US UTF-8"
   },
   "en_US.UTF-8": {
      "UTF-8": 0       // allows "en_US.UTF-8 UTF-8"
   }
}

Note that the decision as to whether to generate a short-form alias is
based on the (now meaningful) value of the innermost key, duly ensuring
that there are no backward-compatibility breaks in that regard.

The mkconfig utility has been updated so as to refrain from
incorporating the UTF-8 charmap into the first field of each of the
entries that it generates.

Finally, The man pages have yet to be modified so as to precisely
document the validation behaviour. They shall be attended to by a
forthcoming commit.

See-also: 588e3ede89a6b630420839e704cf975a28f0d981
See-also: d1d1854a402b02b23b08725dac069984506f1f96
Signed-off-by: Kerin Millar <kfm@plushkava.net>